Move hive metastores' tests to specific packages
Hive tests have packages for thrift, file, glue metastores' tests. Move
the tests from outside into these packages.
